feat: configuração do docker
