+++ backend/README
@@ -29,5 +29,21 @@
 ## O que é o cors?
 CORS significa Cross-Origin Resource Sharing (Compartilhamento de Recursos de Origem Cruzada). Por segurança, os navegadores bloqueiam requisições feitas de um domínio (ex: meu-site.com) para outro (ex: api.meu-site.com), a menos que o servidor da API dê uma "autorização" explícita.
 
+## Importante (Organização de pastas):
+- Controllers/: recebe a requisição;
+- Repository/: faz as queries (Knex);
+- Services/: Regras de negócio;
+- Types/: ou arquivo interface.ts: Define o formato do dado;
 
+## Sobre o docker-compose:
+O Gemini disse:
+O docker-compose.yml é essencial para que o seu professor ou qualquer outro desenvolvedor consiga rodar o seu projeto (API + Banco de Dados) com apenas um comando, sem precisar instalar o PostgreSQL manualmente na máquina.
 
+  ## -- Para subir tudo: No terminal, na raiz do projeto, digite:
+  docker-compose up -d
+
+  ## -- Para ver se está rodando:
+  docker ps
+
+  ## -- Para derrubar tudo:
+  docker-compose down
