+++ backend/src/modules/avaliacoes-vinculo/controller/avaliacao-controller.ts
@@ -1,0 +1,1 @@
+import type { Avaliacao } from '../models/interface.js';
