+++ backend/src/modules/avaliacoes-vinculo/services/avaliacao-service.ts
@@ -1,0 +1,1 @@
+import type { Avaliacao } from '../models/interface.js';
